pyproject.toml
[project]
name = "evalpoint-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu==1.10",
    "jupyter>=1.1.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.19",
    "langchainhub>=0.1.21",
    "langgraph>=0.4.8",
    "matplotlib>=3.10.3",
    "openai>=1.84.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "ragas>=0.2.15",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "torchvision>=0.22.0",
    "tqdm>=4.67.1",
]
